+++ packages/codemirror-block-editor/src/index.ts
@@ -1,37 +1,76 @@
 import { basicSetup, EditorState, EditorView } from "@codemirror/basic-setup";
 import { Extension, StateEffect, StateField } from "@codemirror/state";
 import { Decoration, DecorationSet, keymap } from "@codemirror/view";
-import { vim } from "@replit/codemirror-vim";
+import { TreeCursor } from "@lezer/common";
+import { Block } from "./block-document";
 
 const baseTheme = EditorView.baseTheme({
   ".cm-block-indentation": { paddingLeft: "3rem" },
 });
 
-const blockIndendationDecoration = Decoration.line({
-  attributes: { class: "cm-block-indentation" },
-});
+const increaseBlockIndentationEffect = StateEffect.define<number>();
+const decreaseBlockIndentationEffect = StateEffect.define<number>();
 
-const increaseBlockIndentation = StateEffect.define<number>();
-const decreaseBlockIndentation = StateEffect.define<number>();
+// const blockState = StateField.define<Block[]>({
+//   create(state: EditorState) {
+//     console.log("creating");
+//     return [];
+//   },
+//   update(blocks, transaction) {
+//     // transaction.changes.iterChanges((fromA, toA, fromB, toB, text) => {
+//     //   console.log("change", fromA, toA, fromB, toB, text.toJSON());
+//     // });
+//     console.log(
+//       transaction.state.field(blockIndentationDecorationsField, false)
+//     );
 
-const blockIndentationField = StateField.define<DecorationSet>({
+//     return blocks;
+//   },
+// });
+
+const blockIndentationDecorationsField = StateField.define<DecorationSet>({
   create() {
     return Decoration.none;
   },
   update(indentations, transaction) {
+    const findLevelOfLine = (line: number) => {
+      let level = 0;
+      indentations.between(line, line, (_, __, value) => {
+        // find current indentation of block
+        level = Number.parseInt(value.spec.attributes.blockLevel);
+        return false;
+      });
+      return level;
+    };
+    const updateBlockLevelOfLine = (line: number, level: number) => {
+      let add = [];
+      if (level > 0) {
+        add.push(
+          Decoration.line({
+            attributes: {
+              style: `padding-left: ${level + 1}rem`,
+              blockLevel: `${level}`,
+            },
+          }).range(line)
+        );
+      }
+      indentations = indentations.update({
+        filter: (from) => from !== line,
+        add,
+      });
+      console.log("set block indentation for line", line, " to level ", level);
+    };
     indentations = indentations.map(transaction.changes);
     for (const effect of transaction.effects) {
-      if (effect.is(increaseBlockIndentation)) {
+      if (effect.is(increaseBlockIndentationEffect)) {
         console.log("increase block indentation for line", effect.value);
-        indentations = indentations.update({
-          add: [blockIndendationDecoration.range(effect.value)],
-        });
+        const level = findLevelOfLine(effect.value) + 1;
+        updateBlockLevelOfLine(effect.value, level);
       }
-      if (effect.is(decreaseBlockIndentation)) {
+      if (effect.is(decreaseBlockIndentationEffect)) {
         console.log("decrease block indentation for line", effect.value);
-        indentations = indentations.update({
-          filter: (from) => from !== effect.value,
-        });
+        const level = Math.max(findLevelOfLine(effect.value) - 1, 0);
+        updateBlockLevelOfLine(effect.value, level);
       }
     }
     return indentations;
@@ -39,15 +78,15 @@ const blockIndentationField = StateField.define<DecorationSet>({
   provide: (f) => EditorView.decorations.from(f),
 });
 
-export function indentBlock(view: EditorView) {
+export function indentBlockCommand(view: EditorView) {
   let effects: StateEffect<unknown>[] = view.state.selection.ranges.map(
     (range) =>
-      increaseBlockIndentation.of(view.state.doc.lineAt(range.from).from)
+      increaseBlockIndentationEffect.of(view.state.doc.lineAt(range.from).from)
   );
   if (!effects.length) return false;
-  if (!view.state.field(blockIndentationField, false)) {
+  if (!view.state.field(blockIndentationDecorationsField, false)) {
     effects.push(
-      StateEffect.appendConfig.of([blockIndentationField, baseTheme])
+      StateEffect.appendConfig.of([blockIndentationDecorationsField, baseTheme])
     );
   }
   view.dispatch({ effects });
@@ -55,15 +94,15 @@ export function indentBlock(view: EditorView) {
   return true;
 }
 
-export function unindentBlock(view: EditorView) {
+export function unindentBlockCommand(view: EditorView) {
   let effects: StateEffect<unknown>[] = view.state.selection.ranges.map(
     (range) =>
-      decreaseBlockIndentation.of(view.state.doc.lineAt(range.from).from)
+      decreaseBlockIndentationEffect.of(view.state.doc.lineAt(range.from).from)
   );
   if (!effects.length) return false;
-  if (!view.state.field(blockIndentationField, false)) {
+  if (!view.state.field(blockIndentationDecorationsField, false)) {
     effects.push(
-      StateEffect.appendConfig.of([blockIndentationField, baseTheme])
+      StateEffect.appendConfig.of([blockIndentationDecorationsField, baseTheme])
     );
   }
   view.dispatch({ effects });
@@ -76,12 +115,12 @@ const indendationKeymap = keymap.of([
     preventDefault: true,
     run: (view: EditorView) => {
       console.log("Tab");
-      indentBlock(view);
+      indentBlockCommand(view);
       return true;
     },
     shift: (view: EditorView) => {
       console.log("Shift-Tab");
-      unindentBlock(view);
+      unindentBlockCommand(view);
       return true;
     },
   },
@@ -99,6 +138,7 @@ const initialState = EditorState.create({
   extensions: [
     //vim(),
     blockIndentationExtension(),
+    // blockState,
     basicSetup,
   ],
 });