+++ packages/codemirror-block-editor/CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 0.1.0
 
 - Basic indentation with Tab and Shift-Tab
+- Increase and decrease levels of lines
 - Initial version